requirements.txt
openai==2.16.0
python-dotenv==1.2.1
pydantic-settings==2.12.0
pydantic==2.12.5
uvicorn==0.40.0
fastapi==0.128.0
python-jose==3.5.0
pytest==9.0.2
flake8==7.3.0
black==25.12.0
pyyaml==6.0.3
pytest-asyncio==1.3.0
redis==7.2.0